+++ source/glest_map_editor/map.cpp
@@ -76,8 +76,8 @@ int Map::getStartLocationY(int index) const {
 }
 
 static int get_dist(int delta_x, int delta_y) {
-	float dx = delta_x;
-	float dy = delta_y;
+	float dx = (float) delta_x;
+	float dy = (float)delta_y;
 
 	return static_cast<int>(sqrtf(dx * dx + dy * dy));
 }
@@ -127,7 +127,7 @@ void Map::pirateChangeHeight(int x, int y, int height, int radius) {
 	// If the radius is 1 don't bother doing any calculations
 	if (radius == 1) {
 		if(inside(x, y)){
-			cells[x][y].height = goalAlt;
+			cells[x][y].height = (float)goalAlt;
 		}
 		return;
 	}
